Factor out cartesian parts.
Just copied text over to new file and added required imports on top.
In __init__.py, added import from new .cartesian at bottom to avoid
circular imports.
